Backend/requirements.txt
a2wsgi==1.5.0
arrow==1.2.2
azure-common==1.1.28
azure-core==1.24.2
azure-mgmt-core==1.3.1
azure-mgmt-redis==13.1.0
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==7.1.2
cookiecutter==1.7.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.68.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
isodate==0.6.1
manage-fastapi==1.1.1
mysql-connector-python==8.0.30
pydantic==1.9.1
pymssql==2.2.5
PyMySQL==1.0.2
pyodbc==4.0.34
python-dateutil==2.8.2
python-dotenv==0.20.0
SQLAlchemy==1.4.39
starlette==0.14.2
text-unidecode==1.3
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.12.2
uvloop==0.16.0
wfastcgi==3.0.0